net: wireless: bcmdhd: Fix up the BRCM wifi DHD code
Security Vulnerability fix for memory overflow wifi driver
function wl_cfgvendor_rtt_set_config. In the current fix added
check to validate if the target_cnt is valid or not if it is not valid
then parse error. Since target_cnt can be controlled by user netlink
input which needs to validated at the DHD level.

Signed-off-by: Sudhir Kohalli <sudhir.kohalli@broadcom.com>
Bug: 32219255
Change-Id: I5cf771c60a6ae8019e5e36571197e2849c572b40